requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
opencv-python-headless==4.7.0.72
requests==2.28.2
ultralytics==8.0.196
Werkzeug==2.2.3
numpy==1.23.5
torch==2.0.1
pillow==9.5.0
geopy==2.3.0
python-telegram-bot==13.7 